Move the source file to .Net Core project.
