+++ Vsxmd.Legacy/Vsxmd.Legacy.csproj
@@ -68,27 +68,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Converter.cs" />
-    <Compile Include="IConverter.cs" />
-    <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TableOfContents.cs" />
-    <Compile Include="Units\AssemblyUnit.cs" />
-    <Compile Include="Units\BaseUnit.cs" />
-    <Compile Include="Units\ExampleUnit.cs" />
-    <Compile Include="Units\ExceptionUnit.cs" />
-    <Compile Include="Units\Extensions.cs" />
-    <Compile Include="Units\IUnit.cs" />
-    <Compile Include="Units\MemberKind.cs" />
-    <Compile Include="Units\MemberName.cs" />
-    <Compile Include="Units\MemberUnit.cs" />
-    <Compile Include="Units\ParamUnit.cs" />
-    <Compile Include="Units\PermissionUnit.cs" />
-    <Compile Include="Units\RemarksUnit.cs" />
-    <Compile Include="Units\ReturnsUnit.cs" />
-    <Compile Include="Units\SeealsoUnit.cs" />
-    <Compile Include="Units\SummaryUnit.cs" />
-    <Compile Include="Units\TypeparamUnit.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config">
