+++ Vsxmd/Converter.cs
@@ -59,10 +59,6 @@ namespace Vsxmd
             // table of contents
             var tableOfContents = new TableOfContents(memberUnits);
 
-            //return new IUnit[] { tableOfContents }
-            //    .Concat(new[] { assemblyUnit })
-            //    .Concat(memberUnits);
-
             return new IUnit[] { assemblyUnit }
                 .Concat(new[] { tableOfContents })
                 .Concat(memberUnits);
